Added support for saving and reading settings to eeprom
Added support for retreiving the detector name - can be done as two parts to present nicely on the display - part one on one line, part two on the one below.
Added support for editing settings without having to reflash the arduino - type "Settings" and a new line before the detector starts up fully.